requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
Django==5.1.2
djangorestframework==3.15.2
exceptiongroup==1.2.0
executing==2.0.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==9.1.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
scipy==1.11.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.12
